executors: move some test files into cmdexec
The tests in sshexec were a bit intertwined between testing (mock) ssh
functionality versus the functionality of the major cli running functions.
This change moves the files that were mostly testing the behavior of
functions (brick/peer) into cmdexec.
