+++ executors/cmdexec/brick_test.go
@@ -7,7 +7,7 @@
 // cases as published by the Free Software Foundation.
 //
 
-package sshexec
+package cmdexec
 
 import (
 	"strings"
@@ -19,25 +19,11 @@ import (
 )
 
 func TestSshExecBrickCreate(t *testing.T) {
-
-	f := NewFakeSsh()
-	defer tests.Patch(&sshNew,
-		func(logger *utils.Logger, user string, file string) (Ssher, error) {
-			return f, nil
-		}).Restore()
-
-	config := &SshConfig{
-		PrivateKeyFile: "xkeyfile",
-		User:           "xuser",
-		Port:           "100",
-		CLICommandConfig: CLICommandConfig{
-			Fstab: "/my/fstab",
-		},
-	}
-
-	s, err := NewSshExecutor(config)
+	f := NewCommandFaker()
+	s, err := NewFakeExecutor(f)
 	tests.Assert(t, err == nil)
 	tests.Assert(t, s != nil)
+	s.portStr = "100"
 
 	// Create a Brick
 	b := &executors.BrickRequest{
@@ -105,25 +91,11 @@ func TestSshExecBrickCreate(t *testing.T) {
 }
 
 func TestSshExecBrickCreateWithGid(t *testing.T) {
-
-	f := NewFakeSsh()
-	defer tests.Patch(&sshNew,
-		func(logger *utils.Logger, user string, file string) (Ssher, error) {
-			return f, nil
-		}).Restore()
-
-	config := &SshConfig{
-		PrivateKeyFile: "xkeyfile",
-		User:           "xuser",
-		Port:           "100",
-		CLICommandConfig: CLICommandConfig{
-			Fstab: "/my/fstab",
-		},
-	}
-
-	s, err := NewSshExecutor(config)
+	f := NewCommandFaker()
+	s, err := NewFakeExecutor(f)
 	tests.Assert(t, err == nil)
 	tests.Assert(t, s != nil)
+	s.portStr = "100"
 
 	// Create a Brick
 	b := &executors.BrickRequest{
@@ -202,26 +174,12 @@ func TestSshExecBrickCreateWithGid(t *testing.T) {
 }
 
 func TestSshExecBrickCreateSudo(t *testing.T) {
-
-	f := NewFakeSsh()
-	defer tests.Patch(&sshNew,
-		func(logger *utils.Logger, user string, file string) (Ssher, error) {
-			return f, nil
-		}).Restore()
-
-	config := &SshConfig{
-		PrivateKeyFile: "xkeyfile",
-		User:           "xuser",
-		Port:           "100",
-		CLICommandConfig: CLICommandConfig{
-			Fstab: "/my/fstab",
-			Sudo:  true,
-		},
-	}
-
-	s, err := NewSshExecutor(config)
+	f := NewCommandFaker()
+	s, err := NewFakeExecutor(f)
 	tests.Assert(t, err == nil)
 	tests.Assert(t, s != nil)
+	s.useSudo = true
+	s.portStr = "100"
 
 	// Create a Brick
 	b := &executors.BrickRequest{
@@ -290,25 +248,11 @@ func TestSshExecBrickCreateSudo(t *testing.T) {
 }
 
 func TestSshExecBrickDestroy(t *testing.T) {
-
-	f := NewFakeSsh()
-	defer tests.Patch(&sshNew,
-		func(logger *utils.Logger, user string, file string) (Ssher, error) {
-			return f, nil
-		}).Restore()
-
-	config := &SshConfig{
-		PrivateKeyFile: "xkeyfile",
-		User:           "xuser",
-		Port:           "100",
-		CLICommandConfig: CLICommandConfig{
-			Fstab: "/my/fstab",
-		},
-	}
-
-	s, err := NewSshExecutor(config)
+	f := NewCommandFaker()
+	s, err := NewFakeExecutor(f)
 	tests.Assert(t, err == nil)
 	tests.Assert(t, s != nil)
+	s.portStr = "100"
 
 	// Create a Brick
 	b := &executors.BrickRequest{
